pyproject.toml
[project]
name = "everydoor-plugin-repo"
version = "0.1.0"
description = "Every Door Plugin Repository"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.5.2",
    "flask>=3.1.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "markdown2>=2.5.3",
    "pyyaml>=6.0.2",
    "qrcode>=8.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
]
